Fix incorrect point placement in minibuffer when prefixed.
Users can prefix an ag-command to pass additional arguments.  If so,
all the arguments will be prompted in the minibuffer for user editing
and point is adjusted to the position for user's input.

However, the calculation of the point position does not take the
backslash (after shell-quote-argument) into account, nor the
variable-length of the `:files' argument.

This commit fixes this issue by searching for " -- " and place the
cursor before it rather than subtracting a fixed value.